bugfix: infinite loop while visiting nodes
While the tree optimizer was traversing the AST,
sometimes it entered infite loops due to cycles in
the ADG (actually the AST is not a tree anymore). Fixed.

Actually, upon further investigation, it seems it not
"infinite loop time", but a "very very long time". To sum up,
the optimizer was doing many repeated (and useles) tree
traversal passed. Now they're dramatically reduced and so, the
compiler has been speed up!

Also add tests with a timeout to detect long compiling times.