Add routine to set ATTR to address
This routine, given a pixel address on the screen in HL will
set the ATTR for it.

Also adds an extra entry point to the SET_ATTR routine.

This routine is in charge of setting the ATTR upon an screen
update operation (PRINT, PLOT, DRAW, CIRCLE).

Now it can be used also by drawing primitives instead of the ROM's one.