Bugfix: correctly compute labels for JR
Relative jumps (JR, DJNZ) were computer wrongly when being
using in a line with several instructions. Fixed.

The syntax is now extended, so a LABEL does not need to use colon, and
can also be declared in between instructions. So:

LOOP
  ...

is valid, but also:

ld a, 0 : LOOP : nop : jp LOOP