Add symbol dependency
All Symbol instances now have a set which points which
others it requires and which ones requires the given instance.

An symbol that requires another means that all the parents of the former
will also requires it.

Whenever a symbol "depends" on another (i.e. it needs some attribute of it,
or makes use of it for whatever) this is now referenced in a counter collection.
The counter allows to signal the same reference multiple times, so for example:

LET c = a
LET c = c + a

Can be denoted as "c" requires "a" twice.