Remove unused labels in O2
This update tried hard to remove unused
labels. Specially temporary ones.
User labels are left untouched.

If a several labels point to the same memory address
only the last one prevails, except when it's an user label.

If the last label is a temporary label and any other is
an user label, the latter will prevail.