Do not mark labels as accessed
Labels are not marked as accessed until
the Optimization stage. This will need to be
refactored since the responsibility of deciding
when an ID is accessed, and guess what kind is it
or it's type when referenced in advance will be
moved to the visitors.
The parser will be slowly moving towards parsing-only
tasks.