fix: correctly compute arrays offset within function
This required a revamp of the way the arrays were used.
The array.asm runtime now uses the LBOUND table *always*
unless the array is zero-based in all dimensions.

On each iteration, the index is subtracted the i-th lower-bound.
There is no longer the need to subtract it when pushing it.